+++ lib/internal/bootstrap/node.js
@@ -214,6 +214,16 @@ if (!config.noBrowserGlobals) {
   defineOperation(global, 'setImmediate', timers.setImmediate);
 }
 
+// Set the per-Environment callback that will be called
+// when the TrackingTraceStateObserver updates trace state.
+// Note that when NODE_USE_V8_PLATFORM is true, the observer is
+// attached to the per-process TracingController.
+const { setTraceCategoryStateUpdateHandler } = internalBinding('trace_events');
+const {
+  toggleTraceCategoryState
+} = NativeModule.require('internal/process/per_thread');
+setTraceCategoryStateUpdateHandler(toggleTraceCategoryState);
+
 // process.allowedNodeEnvironmentFlags
 Object.defineProperty(process, 'allowedNodeEnvironmentFlags', {
   get() {
