+++ lib/internal/bootstrap/pre_execution.js
@@ -1,9 +1,6 @@
 'use strict';
 
 const { getOptionValue } = require('internal/options');
-// Lazy load internal/trace_events_async_hooks only if the async_hooks
-// trace event category is enabled.
-let traceEventsAsyncHook;
 
 function prepareMainThreadExecution() {
   // Patch the process object with legacy properties and normalizations
@@ -162,26 +159,9 @@ function initializeReportSignalHandlers() {
 }
 
 function setupTraceCategoryState() {
-  const {
-    asyncHooksEnabledInitial,
-    setTraceCategoryStateUpdateHandler
-  } = internalBinding('trace_events');
-
-  toggleTraceCategoryState(asyncHooksEnabledInitial);
-  setTraceCategoryStateUpdateHandler(toggleTraceCategoryState);
-}
-
-// Dynamically enable/disable the traceEventsAsyncHook
-function toggleTraceCategoryState(asyncHooksEnabled) {
-  if (asyncHooksEnabled) {
-    if (!traceEventsAsyncHook) {
-      traceEventsAsyncHook =
-        require('internal/trace_events_async_hooks').createHook();
-    }
-    traceEventsAsyncHook.enable();
-  } else if (traceEventsAsyncHook) {
-    traceEventsAsyncHook.disable();
-  }
+  const { isTraceCategoryEnabled } = internalBinding('trace_events');
+  const { toggleTraceCategoryState } = require('internal/process/per_thread');
+  toggleTraceCategoryState(isTraceCategoryEnabled('node.async_hooks'));
 }
 
 // In general deprecations are intialized wherever the APIs are implemented,
